Add api/web application
The purpose of this commit is to add 'api' folder to the repository. 
This folder contains the API/web application.

static folder contains the static content of the web application such as 
css or js files.
templates folder contains the html files of the web application.
bf_api.py will be used to create the web app and the API.
Dockerfile will be used to create the docker image for the API/web 
application.
README.md contains the descripion of the API
requirements.txt contains the modules nedeed for the API/web 
application. The Docker file will look for that modules here.