Add models which describe the database
base.py contains the objects engine, Base and Session and also the 
configuration parameters for the database (host, port, user, password, 
databse name)
Each of the other files contains a class that describes a table or a 
view in the database. For example product.py contains Product class 
which describes product table in the database.